add unit tests for assignment 6

+++ Application_programming_in_java/Homeworks/CP125Assignment6/src/main/java/com/scg/beans/HumanResourceManager.java
@@ -98,7 +98,7 @@ public class HumanResourceManager {
      * @param consultant The consultant whose pay rate is to be changed.
      * @param newPayRate The proposed new pay rate.
      */
-    public void adjustPayRate( StaffConsultant consultant, int newPayRate ){
+    public void adjustPayRate(StaffConsultant consultant, double newPayRate ){
         String message = "Pay rate change for: " + consultant.getName().toString() + "; ";
         try {
             consultant.setPayRate(newPayRate);
@@ -203,8 +203,5 @@ public class HumanResourceManager {
     public void removeBenefitListener( BenefitListener listener ){
         benefitListeners.remove(listener);
     }
-
-
-
 
 }
